@@ -6,13 +6,13 @@ class KeyboardVisibilitySubscriber {
   /// Called when a keyboard visibility event occurs
   /// Is only called when the state changes
   /// The [visible] parameter reflects the new visibility
-  final Function(bool visible) onChange;
+  final Function(bool visible)? onChange;
 
   /// Called when the keyboard appears
-  final Function onShow;
+  final Function? onShow;
 
   /// Called when the keyboard closes
-  final Function onHide;
+  final Function? onHide;
 
   /// Constructs a new [KeyboardVisibilitySubscriber]
   KeyboardVisibilitySubscriber({this.onChange, this.onShow, this.onHide});
@@ -24,7 +24,7 @@ class KeyboardVisibilityNotification {
       const EventChannel('github.com/adee42/flutter_keyboard_visibility');
   static Map<int, KeyboardVisibilitySubscriber> _list =
       Map<int, KeyboardVisibilitySubscriber>();
-  static StreamSubscription _keyboardVisibilitySubscription;
+  static StreamSubscription? _keyboardVisibilitySubscription;
   static int _currentIndex = 0;
 
   /// The current state of the keyboard visibility. Can be used without subscribing
@@ -39,19 +39,19 @@ class KeyboardVisibilityNotification {
 
   /// Internal function to handle native code channel communication
   void onKeyboardEvent(dynamic arg) {
-    isKeyboardVisible = (arg as int) == 1;
+    isKeyboardVisible = (arg as int?) == 1;
 
     // send a message to all subscribers notifying them about the new state
     _list.forEach((subscriber, s) {
       try {
         if (s.onChange != null) {
-          s.onChange(isKeyboardVisible);
+          s.onChange!(isKeyboardVisible);
         }
         if ((s.onShow != null) && isKeyboardVisible) {
-          s.onShow();
+          s.onShow!();
         }
         if ((s.onHide != null) && !isKeyboardVisible) {
-          s.onHide();
+          s.onHide!();
         }
       } catch (_) {}
     });
@@ -63,7 +63,7 @@ class KeyboardVisibilityNotification {
   /// [onHide] is called when the keyboard disappears
   /// Returns a subscribing id that can be used to unsubscribe
   int addNewListener(
-      {Function(bool) onChange, Function onShow, Function onHide}) {
+      {Function(bool)? onChange, Function? onShow, Function? onHide}) {
     _list[_currentIndex] = KeyboardVisibilitySubscriber(
         onChange: onChange, onShow: onShow, onHide: onHide);
     return _currentIndex++;
